PDF_Query_using_LangGraph/requirements.txt
ipykernel
python-dotenv
PyPDF2 
langgraph
langchain
langchain-community
langchain-google-genai
faiss-cpu 
streamlit